test_schwimmbad: test manual pool creation, test null timer

+++ tests/test_schwimmbad.py
@@ -31,9 +31,16 @@ def test_dill():
 
 
 @pytest.mark.mpi
-def test_timer():
+@pytest.mark.parametrize("enabled", [False, True])  # False case tests the null timer
+def test_timer(enabled):
+    if enabled:
+        interval = 1.0e-3
+    else:
+        interval = None
+
     sleep_time = 1.0
-    with MPIPool(time_mpi=True, timing_interval=1.0e-3) as pool:
+
+    with MPIPool(time_mpi=enabled, timing_interval=interval) as pool:
         if pool.is_master():
             results = _pool_test(pool)
             # Trigger the "master_serial" timer
@@ -41,39 +48,54 @@ def test_timer():
             results = _pool_test(pool)
             assert all(0 <= x <= 2 for x in results)
 
-    # Need to run tests outside of pool context, but only on master
-    comm = MPI.COMM_WORLD
-    if comm.Get_rank() == 0:
-        timing_file = "mpi_worker_timing.json"
-        assert os.path.exists(timing_file)
+    if enabled:
+        # Need to run tests outside of pool context, but only on master
+        comm = MPI.COMM_WORLD
+        if comm.Get_rank() == 0:
+            timing_file = "mpi_worker_timing.json"
+            assert os.path.exists(timing_file)
 
-        try:
-            with open(timing_file, "r") as f:
-                data = json.load(f)
-        finally:
-            os.remove(timing_file)
+            try:
+                with open(timing_file, "r") as f:
+                    data = json.load(f)
+            finally:
+                os.remove(timing_file)
 
-        assert type(data) == list
-        assert len(data) >= 1
-        assert type(data[0]) == dict
+            assert type(data) == list
+            assert len(data) >= 1
+            assert type(data[0]) == dict
 
-        expected_keys = [
-            "master_serial",
-            "mpi_recv",
-            "compute",
-            "mpi_send",
-            "barrier",
-            "walltime",
-        ]
-        for entry in expected_keys:
-            assert entry in data[0].keys()
+            expected_keys = [
+                "master_serial",
+                "mpi_recv",
+                "compute",
+                "mpi_send",
+                "barrier",
+                "walltime",
+            ]
+            for entry in expected_keys:
+                assert entry in data[0].keys()
 
-        assert sum(step["master_serial"] for step in data) == pytest.approx(
-            sleep_time, abs=1.0e-2
-        )
+            assert sum(step["master_serial"] for step in data) == pytest.approx(
+                sleep_time, abs=1.0e-2
+            )
 
 
 @pytest.mark.mpi
 def test_enabled():
     with MPIPool() as pool:
         assert pool.enabled()
+
+
+@pytest.mark.mpi
+def test_without_context():
+    # Open the pool manually
+    pool = MPIPool()
+
+    results = _pool_test(pool)
+    if pool.is_master():
+        assert all(0 <= x <= 2 for x in results)
+
+    # Close the pool manually
+    pool.close()
+    assert pool.pool_open is False
